requirements.txt
streamlit
faiss-cpu
pypdf
sentence-transformers
numpy
requests